backend/requirements_snapshot.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.34
botocore==1.42.34
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cryptography==46.0.3
curl_cffi==0.13.0
distro==1.9.0
docling-core==2.61.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
exchange_calendars==4.12
fake-useragent==2.2.0
fastapi==0.128.0
frozendict==2.4.7
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
invoke==2.2.1
jiter==0.12.0
jmespath==1.1.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
korean-lunar-calendar==0.3.1
latex2mathml==3.78.1
logistro==2.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
mistralai==1.10.1
multitasking==0.0.12
narwhals==2.15.0
nselib==2.4.2
numpy==2.4.1
openai==2.15.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.5
packaging==26.0
pandas==2.3.3
pandas_market_calendars==5.2.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyluach==2.3.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scipy==1.17.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
toolz==1.1.0
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yfinance==1.1.0
zipp==3.23.0
